makin makin

+++ DAL/users_crud.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException
 from starlette.responses import JSONResponse
 
 from models import User
@@ -17,7 +18,8 @@ def create_user(db: Session, user: CreateUpdateUserSchema):
     _user = User(
         username=user.username,
         email=user.email,
-        hashed_password=user.password
+        hashed_password=user.password,
+        balance = 0
 
     )
     db.add(_user)
@@ -32,12 +34,17 @@ def get_active_users(db: Session):
 
 def remove_user(db: Session, user_id: int):
     _user = get_user_by_id(db=db, user_id=user_id)
+    if not _user :
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(_user)
     db.commit()
     return JSONResponse(content={"detail": "OK"}, status_code=200)
 
+
 def update_user(db: Session, user_id: int, user: CreateUpdateUserSchema):
     _user = get_user_by_id(db=db, user_id=user_id)
+    if not _user :
+        raise HTTPException(status_code=404, detail="User not found")
 
     _user.username = user.username
     _user.email = user.email
